// server/src/payments/domain/enums/FulfillmentTransition.ts
export enum FulfillmentTransition {
  INITIALIZED = 'INITIALIZED',
  ASSIGN_AGENT = 'ASSIGN_AGENT',
  START_DELIVERY = 'START_DELIVERY',
  VERIFY_OTP = 'VERIFY_OTP',
  COMPLETE_DELIVERY = 'COMPLETE_DELIVERY',
  COMPLETE_PICKUP = 'COMPLETE_PICKUP',
  FAIL_DELIVERY = 'FAIL_DELIVERY',
  REGENERATE_OTP = 'REGENERATE_OTP',
  REJECT_AGENT = 'REJECT_AGENT'
}